add cast CLogicInt (COpe)

+++ sakura_core/cmd/COpe.cpp
@@ -34,8 +34,8 @@ COpe::~COpe()
 void COpe::DUMP( )
 {
 	DEBUG_TRACE( L"\t\tm_nOpe                  = [%d]\n", m_nOpe               );
-	DEBUG_TRACE( L"\t\tm_ptCaretPos_PHY_Before = [%d,%d]\n", m_ptCaretPos_PHY_Before.x, m_ptCaretPos_PHY_Before.y   );
-	DEBUG_TRACE( L"\t\tm_ptCaretPos_PHY_After  = [%d,%d]\n", m_ptCaretPos_PHY_After.x, m_ptCaretPos_PHY_After.y   );
+	DEBUG_TRACE( L"\t\tm_ptCaretPos_PHY_Before = [%d,%d]\n", static_cast<int>(m_ptCaretPos_PHY_Before.x), static_cast<int>(m_ptCaretPos_PHY_Before.y) );
+	DEBUG_TRACE( L"\t\tm_ptCaretPos_PHY_After  = [%d,%d]\n", static_cast<int>(m_ptCaretPos_PHY_After.x), static_cast<int>(m_ptCaretPos_PHY_After.y) );
 	return;
 }
 
@@ -43,7 +43,7 @@ void COpe::DUMP( )
 void CDeleteOpe::DUMP( )
 {
 	COpe::DUMP();
-	DEBUG_TRACE( L"\t\tm_ptCaretPos_PHY_To     = [%d,%d]\n", m_ptCaretPos_PHY_To.x, m_ptCaretPos_PHY_To.y );
+	DEBUG_TRACE( L"\t\tm_ptCaretPos_PHY_To     = [%d,%d]\n", static_cast<int>(m_ptCaretPos_PHY_To.x), static_cast<int>(m_ptCaretPos_PHY_To.y) );
 	DEBUG_TRACE( L"\t\tm_cOpeLineData.size         = [%d]\n", m_cOpeLineData.size() );
 	for( size_t i = 0; i < m_cOpeLineData.size(); i++ ){
 		DEBUG_TRACE( L"\t\tm_cOpeLineData[%d].nSeq         = [%d]\n", m_cOpeLineData[i].nSeq );
